+++ src/brasil/gov/agenda/Extensions/Install.py
@@ -6,7 +6,8 @@ def uninstall(portal, reinstall=False):
     if not reinstall:
         tinymce = api.portal.get_tool('portal_tinymce')
         linkable = tinymce.linkable.split()
-        if 'Agenda' in linkable:
-            linkable.remove('Agenda')
-            tinymce.linkable = u'\n'.join(linkable)
-            return 'Agenda removed from linkable types in TinyMCE.'
+        for t in ('Agenda', 'AgendaDiaria', 'Compromisso'):
+            if t in linkable:
+                linkable.remove(t)
+        tinymce.linkable = u'\n'.join(linkable)
+        return 'Package content types are no longer linkable on TinyMCE.'