+++ src/brasil/gov/agenda/upgrades/v4003/__init__.py
@@ -3,14 +3,15 @@ from brasil.gov.agenda.logger import logger
 from plone import api
 
 
-def make_agenda_linkable_on_tinymce(context):
-    """Add Agenda to the list of linkable content types on TinyMCE."""
+def make_content_types_linkable(context):
+    """Add package content types to the list of linkables on TinyMCE."""
     tinymce = api.portal.get_tool('portal_tinymce')
     linkable = tinymce.linkable.split()
-    if 'Agenda' not in linkable:
-        linkable.append('Agenda')
-        tinymce.linkable = u'\n'.join(linkable)
-        logger.info('Agenda added to linkable types in TinyMCE.')
+    for t in ('Agenda', 'AgendaDiaria', 'Compromisso'):
+        if t not in linkable:
+            linkable.append(t)
+            tinymce.linkable = u'\n'.join(linkable)
+            logger.info('{} added to linkable types in TinyMCE.'.format(t))
 
 
 def update_effective_date_index(context):